more refactoring and such things
