+++ Assets/materials/yuukabbage.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0b2dd9dc5bda1ab4f95e843d29a193f5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
