+++ Assets/prefabs/Props/BoostPad.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 045b9efa6f959d048b354389724a04dc
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
