--- Assets/scripts/Car/PlayerCarColors.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bf77673f73d573c489fdc8dfb2a2c314
