+++ Assets/scripts/Enviroment/BoostPad.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 727f8d83eb803844ab42ee7d1615abb8
