+++ Assets/scripts/UI/SpeedMeter.cs
@@ -1,18 +1,16 @@
 using UnityEngine;
-using UnityEngine.UI;
 
 public class SpeedMeter : MonoBehaviour
 {
     public Rigidbody target;
 
-    [Header("float valuet")]
     public float maxSpeed = 240.0f;
     public float minSpeedArrowAngle = 20.0f;
     public float maxSpeedArrowAngle = -200.0f;
 
-    private float speed;
+    private float speed => target.linearVelocity.magnitude * 3.6f;
 
-    [Header("UI")] //kiitos leo's leikkimaa, very cool
+    [Header("UI")]
     public RectTransform arrow;
 
     private void Start()
@@ -22,7 +20,6 @@ public class SpeedMeter : MonoBehaviour
 
     private void Update()
     {
-        speed = target.linearVelocity.magnitude * 3.6f;
-        if (arrow != null) arrow.localEulerAngles = new Vector3(0, 0, Mathf.Lerp(minSpeedArrowAngle, maxSpeedArrowAngle, speed / maxSpeed));
+        if (arrow != null) arrow.localRotation = Quaternion.Euler(0, 0, Mathf.Lerp(minSpeedArrowAngle, maxSpeedArrowAngle, speed / maxSpeed));
     }
 }
