--- Assets/terrain/TerrainData_66c8ef4a-f56d-4998-bd9f-fb5f87f808dd.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d160d42b9dccee4468701a5dc5b3b219
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
