--- Assets/terrain/TerrainData_6d7b2668-d171-4972-b966-ec20a564aefd.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d075c69b2f8e7f04aa9bb774ada7b32c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
