--- Assets/terrain/scenes/New Terrain 8.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 31fe43f0314b28e4db399c439ef7fdd9
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
