--- Assets/terrain/scenes/New Terrain.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c90896d76098930469138556bea931ce
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
