--- Assets/terrain/scenes/juud1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ee7801f836c5daa41a268636d9ab50d1
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
