--- Assets/terrain/scenes/juud2.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c4dd730d808b97640941f765eb733ee5
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
