--- Assets/terrain/scenes/menu.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3b9ee71cc6117a145918fdf4a18db56c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
