--- Assets/terrain/scenes/mountain.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cf61274819e422244a979383468e1b4c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
